Add authenticated /api/json routes for Library.Author
We couldn't add these before as we didn't have a way to sign in a user
(and get their bearer token) via the API. The tests might have passed,
as they internally set the actor on the test post/patch/delete
functions; but we wouldn't be any use for the routes.

As a workaround for duplicate imports between `Phoenix.ConnTest` and
`AshJsonApi.Test`, we don't import `get/3` and `delete/3` from the
former in the `ConnCase`. As we aren't doing direct integration tests
(without `PhoenixTest`) this should be fine.